+++ activejob/test/cases/exceptions_test.rb
@@ -58,8 +58,8 @@ class ExceptionsTest < ActiveJob::TestCase
 
   test "custom handling of job that exceeds retry attempts" do
     perform_enqueued_jobs do
-      RetryJob.perform_later "CustomCatchError", 6
-      assert_equal "Dealt with a job that failed to retry in a custom way", JobBuffer.last_value
+      RetryJob.perform_later 'CustomCatchError', 6
+      assert_equal "Dealt with a job that failed to retry in a custom way after 6 attempts", JobBuffer.last_value
     end
   end